# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SSC Vocabulary Master API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vocabulary:
    get:
      operationId: listVocabulary
      tags: [vocabulary]
      summary: List vocabulary items with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: alphabet
          in: query
          schema:
            type: string
        - name: topic
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Vocabulary list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VocabularyPage"

    post:
      operationId: createVocabulary
      tags: [vocabulary]
      summary: Create vocabulary item (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VocabularyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VocabularyItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /vocabulary/random:
    get:
      operationId: getRandomWords
      tags: [vocabulary]
      summary: Get random vocabulary words (word of the day / quick quiz)
      parameters:
        - name: count
          in: query
          schema:
            type: integer
            default: 5
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Random words
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VocabularyItem"
  /vocabulary/{id}:
    get:
      operationId: getVocabularyItem
      tags: [vocabulary]
      summary: Get single vocabulary item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vocabulary item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VocabularyItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateVocabulary
      tags: [vocabulary]
      summary: Update vocabulary item (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VocabularyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VocabularyItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteVocabulary
      tags: [vocabulary]
      summary: Delete vocabulary item (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Tests ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ─── Auth Schemas ──────────────────────────────────────────────────────
...
    VocabularyItem:
      type: object
      required: [id, word, meaning, category, difficulty, alphabet]
      properties:
        id:
          type: integer
        word:
          type: string
        meaning:
          type: string
        hindiMeaning:
          type: ["string", "null"]
        exampleSentence:
          type: ["string", "null"]
        synonyms:
          type: array
          items:
            type: string
        antonyms:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        category:
          type: string
        alphabet:
          type: string
        topics:
          type: array
          items:
            type: string
        examRefs:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    VocabularyInput:
      type: object
      required: [word, meaning, category, difficulty, alphabet]
      properties:
        word:
          type: string
        meaning:
          type: string
        hindiMeaning:
          type: string
        exampleSentence:
          type: string
        synonyms:
          type: array
          items:
            type: string
        antonyms:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        category:
          type: string
        alphabet:
          type: string
        topics:
          type: array
          items:
            type: string
        examRefs:
          type: array
          items:
            type: string
    VocabularyUpdate:
      type: object
      properties:
        word:
          type: string
        meaning:
          type: string
        hindiMeaning:
          type: string
        exampleSentence:
          type: string
        synonyms:
          type: array
          items:
            type: string
        antonyms:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        category:
          type: string
        alphabet:
          type: string
        topics:
          type: array
          items:
            type: string
        examRefs:
          type: array
          items:
            type: string
        isActive:
          type: boolean
    VocabularyPage:
      type: object
      required: [items, total, page, limit, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/VocabularyItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
...